backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
httpx
pydantic
python-jose[cryptography]